+++ packages/ui/src/index.ts
@@ -6,3 +6,4 @@ export * from "./navigation.js";
 export * from "./select.js";
 export * from "./layout.js";
 export * from "./toast.js";
+export * from "./form.js";